create packages mian and models and move classes to each package
